refactor(asta): extract shared marshal-out helpers into _marshal.py
Move the genuinely-shared marshal-out helpers out of ingest.py (the Paper Finder
module) into a neutral wheeler/integrations/asta/_marshal.py: ImportReport,
_link_once, _edge_exists, _find_paper_by_corpus_id, _paper_exists, _find_execution,
and the corpus_id index load/save. theorizer.py, semantic_scholar.py, and
artifacts.py now import them from one place instead of reaching into ingest.py.

Pure refactor, no behavior change. This is the rule-of-three extraction of the
genuinely shared ~20%; no ToolContract / field-map DSL (YAGNI, as the first
adversarial review flagged). Each adapter keeps its own service tag, parser, and
type-specific dedupe indices.

Full suite: 1816 passed, 2 skipped. ruff + mypy clean.
